requirements.txt
discord.py
requests
python-decouple
youtube-search-python
youtube-dl
